pyproject.toml
[project]
name = "gemini-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=0.1.0",
    "pydantic>=2.12.4",
    "temporalio>=1.19.0",
]
